+++ gear/identifier_lookup/src/python/identifier_app/run.py
@@ -64,7 +64,8 @@ class IdentifierLookupVisitor(GearExecutionEnvironment):
     def __init__(self, *, client: ClientWrapper, admin_id: str,
                  file_input: InputFileWrapper,
                  identifiers_mode: IdentifiersMode, date_field: str,
-                 direction: Literal['nacc', 'center'], gear_name: str):
+                 direction: Literal['nacc', 'center'], gear_name: str,
+                 preserve_case: str):
         super().__init__(client=client)
         self.__admin_id = admin_id
         self.__file_input = file_input
@@ -72,6 +73,7 @@ class IdentifierLookupVisitor(GearExecutionEnvironment):
         self.__direction: Literal['nacc', 'center'] = direction
         self.__date_field = date_field
         self.__gear_name = gear_name
+        self.__preserve_case = preserve_case
 
     @classmethod
     def create(
@@ -97,6 +99,7 @@ class IdentifierLookupVisitor(GearExecutionEnvironment):
         admin_id = context.config.get("admin_group", "nacc")
         mode = context.config.get("database_mode", "prod")
         direction = context.config.get("direction", "nacc")
+        preserve_case = context.config.get("preserve_case", False)
 
         date_field = (context.config.get("date_field",
                                          FieldNames.DATE_COLUMN)).lower()
@@ -109,7 +112,8 @@ class IdentifierLookupVisitor(GearExecutionEnvironment):
                                        file_input=file_input,
                                        identifiers_mode=mode,
                                        date_field=date_field,
-                                       direction=direction)
+                                       direction=direction,
+                                       preserve_case=preserve_case)
 
     def __build_naccid_lookup(self, *, file_id: str,
                               identifiers_repo: IdentifierRepository,
@@ -155,7 +159,8 @@ class IdentifierLookupVisitor(GearExecutionEnvironment):
                                    date_field=self.__date_field,
                                    gear_name=self.__gear_name,
                                    project=ProjectAdaptor(project=project,
-                                                          proxy=self.proxy))
+                                                          proxy=self.proxy),
+                                   preserve_case=self.__preserve_case)
 
     def __build_center_lookup(self, *, identifiers_repo: IdentifierRepository,
                               output_file: TextIO,
@@ -163,7 +168,8 @@ class IdentifierLookupVisitor(GearExecutionEnvironment):
 
         return CenterLookupVisitor(identifiers_repo=identifiers_repo,
                                    output_file=output_file,
-                                   error_writer=error_writer)
+                                   error_writer=error_writer,
+                                   preserve_case=self.__preserve_case)
 
     def run(self, context: GearToolkitContext):
         """Runs the identifier lookup app.