+++ docs/identifier_lookup/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this gear are documented in this file.
 
 ## 1.0.9
 * Fixes bug where `CenterLookupVisitor.visit_row` was setting the keys to lowercase, causing `writer.write()` to fail with case-sensitive headers
+* Adds `preserve_case` configuration value (default `false`) to allow preserving case of header keys in output file
 
 ## 1.0.8
 * Fixes CSV line number in error reports (exclude header row)